correctly create tarball wheelhouse
