+++ msgpack-cxx/pyproject.toml
@@ -13,8 +13,8 @@ build-backend = "setuptools.build_meta"
 name = "msgpack-cxx"
 authors = [{ name = "Alliander Dynamic Grid Calculation", email = "dynamic.grid.calculation@alliander.com" }]
 description = "Repackaging of msgpack-cxx distributed via PyPI"
-readme = "README.md"
-license = { text = "MPL-2.0" }
+readme = "README"
+license = { text = "BSL-1.0" }
 classifiers = [
     "Programming Language :: C++",
     "Development Status :: 5 - Production/Stable",
@@ -45,8 +45,7 @@ Documentation = "https://github.com/msgpack/msgpack-c/tree/cpp_master"
 Mailing-list = "https://lists.lfenergy.org/g/powergridmodel"
 
 [tool.setuptools.packages.find]
-where = ["msgpack-cxx"]  # Generated by automation
-namespaces = false
+where = ['']
 
 [tool.setuptools.dynamic]
 version = {file = "PYPI_VERSION"}
